# frontend/shared/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Physionexs API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/clinic/exercises": {
      "get": {
        "tags": [
          "exercises"
        ],
        "summary": "Clinic Library",
        "description": "Everything this clinic can prescribe: published public exercises + its own.",
        "operationId": "clinic_library_clinic_exercises_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 80
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "name": "body_region",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Body Region"
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 40,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "X-Clinic-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "X-Clinic-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExercisePage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "exercises"
        ],
        "summary": "Create Clinic Exercise",
        "operationId": "create_clinic_exercise_clinic_exercises_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-Clinic-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "X-Clinic-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExerciseIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/clinic/exercises/regions": {
      "get": {
        "tags": [
          "exercises"
        ],
        "summary": "Regions",
        "operationId": "regions_clinic_exercises_regions_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-Clinic-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "X-Clinic-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "title": "Response Regions Clinic Exercises Regions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/clinic/exercises/{exercise_id}/submit": {
      "post": {
        "tags": [
          "exercises"
        ],
        "summary": "Submit For Review",
        "description": "Offer a clinic exercise to the public library. It stays usable in the clinic meanwhile.",
        "operationId": "submit_for_review_clinic_exercises__exercise_id__submit_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "exercise_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Exercise Id"
            }
          },
          {
            "name": "X-Clinic-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "X-Clinic-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/exercises": {
      "get": {
        "tags": [
          "exercises"
        ],
        "summary": "Admin Library",
        "operationId": "admin_library_admin_exercises_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ExerciseStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 80
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "name": "body_region",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Body Region"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExercisePage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "exercises"
        ],
        "summary": "Admin Create",
        "operationId": "admin_create_admin_exercises_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExerciseIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/exercises/{exercise_id}": {
      "get": {
        "tags": [
          "exercises"
        ],
        "summary": "Admin Get",
        "operationId": "admin_get_admin_exercises__exercise_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "exercise_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Exercise Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "exercises"
        ],
        "summary": "Admin Update",
        "operationId": "admin_update_admin_exercises__exercise_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "exercise_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Exercise Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExerciseIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/exercises/{exercise_id}/publish": {
      "post": {
        "tags": [
          "exercises"
        ],
        "summary": "Admin Publish",
        "operationId": "admin_publish_admin_exercises__exercise_id__publish_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "exercise_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Exercise Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewDecisionIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/exercises/{exercise_id}/reject": {
      "post": {
        "tags": [
          "exercises"
        ],
        "summary": "Admin Reject",
        "operationId": "admin_reject_admin_exercises__exercise_id__reject_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "exercise_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Exercise Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewDecisionIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DoseUnit": {
        "type": "string",
        "enum": [
          "reps",
          "seconds"
        ],
        "title": "DoseUnit"
      },
      "ExerciseCategory": {
        "type": "string",
        "enum": [
          "strength",
          "mobility",
          "stretch",
          "balance",
          "proprioception",
          "endurance",
          "breathing",
          "functional"
        ],
        "title": "ExerciseCategory"
      },
      "ExerciseIn": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 160,
            "minLength": 3,
            "title": "Name"
          },
          "body_region": {
            "type": "string",
            "maxLength": 40,
            "minLength": 2,
            "title": "Body Region"
          },
          "category": {
            "$ref": "#/components/schemas/ExerciseCategory"
          },
          "position": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExercisePosition"
              },
              {
                "type": "null"
              }
            ]
          },
          "equipment": {
            "items": {
              "type": "string",
              "maxLength": 40
            },
            "type": "array",
            "maxItems": 8,
            "title": "Equipment"
          },
          "difficulty": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Difficulty",
            "default": 1
          },
          "dose_unit": {
            "$ref": "#/components/schemas/DoseUnit",
            "default": "reps"
          },
          "default_sets": {
            "type": "integer",
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Default Sets",
            "default": 3
          },
          "default_reps": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 200.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Reps",
            "default": 10
          },
          "default_hold_seconds": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 600.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Hold Seconds"
          },
          "default_rest_seconds": {
            "type": "integer",
            "maximum": 600.0,
            "minimum": 0.0,
            "title": "Default Rest Seconds",
            "default": 30
          },
          "steps": {
            "items": {
              "type": "string",
              "maxLength": 400,
              "minLength": 2
            },
            "type": "array",
            "maxItems": 15,
            "minItems": 1,
            "title": "Steps"
          },
          "cues": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 5000
              },
              {
                "type": "null"
              }
            ],
            "title": "Cues"
          },
          "common_mistakes": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 5000
              },
              {
                "type": "null"
              }
            ],
            "title": "Common Mistakes"
          },
          "precautions": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 5000
              },
              {
                "type": "null"
              }
            ],
            "title": "Precautions"
          },
          "contraindications": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 5000
              },
              {
                "type": "null"
              }
            ],
            "title": "Contraindications"
          },
          "conditions": {
            "items": {
              "type": "string",
              "maxLength": 80
            },
            "type": "array",
            "maxItems": 15,
            "title": "Conditions"
          }
        },
        "type": "object",
        "required": [
          "name",
          "body_region",
          "category",
          "steps"
        ],
        "title": "ExerciseIn"
      },
...
      "ExercisePage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ExerciseOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "ExercisePage"
      },
      "ExercisePosition": {
        "type": "string",
        "enum": [
          "standing",
          "sitting",
          "supine",
          "prone",
          "side_lying",
          "quadruped",
          "kneeling"
        ],
        "title": "ExercisePosition"
      },
      "ExerciseStatus": {
        "type": "string",
        "enum": [
          "draft",
          "in_review",
          "published",
          "rejected"
        ],
        "title": "ExerciseStatus"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReviewDecisionIn": {
        "properties": {
          "note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "type": "object",
        "title": "ReviewDecisionIn"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}